Make Pathom error visible to :remote-error? etc
When a resolver throws an error, `:com.wsscode.pathom.core/errors` (a map from attribute to `#:com.fulcrologic.rad.pathom{:errors {:message ..., :data ...}`) is in the body of the response. This change makes it visible to the rest of the code, such as a custom `:remote-error?` (under (-> result :body ::p/errors)).