Convert html-date of "" to nil date object
These functions are used in fulcro-rad-semantic-ui in date pickers. When
using the "clear" button on date picker, the value in the field becomes "".

"" represents an empty value in date picker, so it would make sense that these
html-date-> parsing functions should be able to handle "" and convert it to nil.

Currently they throw a parsing error.