HTML5 History Improvement
The `back!` operation can now auto-detect that there is nowhere
to go (e.g. the app loaded on the page you're on) and instead send
the user to a default route. The back operation will now NEVER exit
the application.

To use the default route you MUST create your HTML5History with
`new-html5-history` and include BOTH the `:default-route` and
`:app` options.