Remove :icons docs, mention rendering plugin ops
As discussed in Slack, the semantics of using :label with an :icon is complicated. For example what if you want only icon, or icon and label, or icon with a hidden, aria-friendly label? So let's not funnel more users to use these and rather instead mention that the rendering plugin may offer better customization options.