Generated resolvers symbol shouldn'tstart with : (#112)
Resolver generated from resolvers that have ao/pathom3-resolve should have their name be a valid, namespaced symbol. 

The old code produced a non-namespaced symbol from a string such as ":order/summary-resolver" which is possible but likely not according to the rules for naming symbols, and it is confusing because it prints like a namespaced keyword but is not.